Update nzbhydra.sh
$(os_codename) failed because it should be => $(_os_codename) and it failed to find package $java because "" should be used instead of ''. 

Nginx configuration fails on recent Ubuntu installations because nzbhydra2 must be run via the Python wrapper which generates the config file successfully, otherwise sed in the nginx config script fails. This script uses "nzbhydra2wrapperPy3.py" for the service. 

If it gets stuck at the website loading, the machine should be restarted and it will work.